id na db é long

+++ app/src/main/java/com/example/myapplication/domain/AbaDeNotasImagemNota.kt
@@ -1,11 +1,10 @@
 package com.example.myapplication.domain
 
 import androidx.room.Entity
-import androidx.room.PrimaryKey
 
 @Entity(primaryKeys = ["idNota","idAba"])
 data class AbaDeNotasImagemNota(
-    val idNota: Int,
+    val idNota: Long,
     val idAba:Int
 ) {
